+++ njunmt/ensemble_experiment.py
@@ -13,18 +13,19 @@
 # limitations under the License.
 """ Define an experiment for ensemble. """
 import time
+
 import tensorflow as tf
 
-from njunmt.nmt_experiment import Experiment
-from njunmt.nmt_experiment import InferExperiment
 from njunmt.data.dataset import Dataset
 from njunmt.data.text_inputter import TextLineInputter
 from njunmt.data.vocab import Vocab
+from njunmt.inference.decode import infer
+from njunmt.models.model_builder import model_fn_ensemble
+from njunmt.nmt_experiment import Experiment
+from njunmt.nmt_experiment import InferExperiment
 from njunmt.utils.configurable import parse_params
 from njunmt.utils.configurable import print_params
 from njunmt.utils.metrics import multi_bleu_score
-from njunmt.utils.model_builder import model_fn_ensemble
-from njunmt.inference.decode import infer
 
 
 class EnsembleExperiment(Experiment):